+++ src/Sensorario/Resources/Exceptions/EmailException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sensorario\Resources\Exceptions;
+
+use Exception;
+
+class EmailException extends Exception { }