+++ src/Sensorario/Resources/Exceptions/PropertyNotSetException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sensorario\Resources\Exceptions;
+
+use Exception;
+
+class PropertyNotSetException extends Exception { }