+++ src/Sensorario/Resources/Exceptions/UndefinedMethodException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sensorario\Resources\Exceptions;
+
+use Exception;
+
+class UndefinedMethodException extends Exception { }